DDMORE-1548: FIS' now doesn't throw a wobbly if can't parse MDL file.
MdlArchiveCreator is called by the MDL->PharmML conversion, and attempts
to extract the data file reference from the MDL so that the referenced
data file can be included in the archive to be passed to the Converter
Toolbox Service. If the MDL file is unparseable then this throws an
exception which is not handled gracefully from the end-user's point of
view, they don't know what went wrong. Instead, catching any
ParseException and continuing to create the archive without any data
file, means that conversion can continue, and the errors in the MDL file
can get reported back to the user via the normal channels.